Refactor math operators prior to primal forumulation package
