+++ demo/dolfin/poisson/hdg_poisson_static_condensation.py
@@ -7,6 +7,7 @@ from dolfin import (
 
 import dolfin_dg.dolfin.hdg_newton
 import dolfin_dg.hdg_form
+import dolfin_dg.math
 
 poly_o = 2
 n_eles = [8, 16, 32, 64]
@@ -58,7 +59,7 @@ for run_no, n_ele in enumerate(n_eles):
     F = inner(F_v(u, grad(u)), grad(v)) * dx
 
     sigma = alpha / h
-    G = dolfin_dg.homogeneity_tensor(F_v, u)
+    G = dolfin_dg.math.homogeneity_tensor(F_v, u)
     hdg_term = dolfin_dg.hdg_form.HDGClassicalSecondOrder(
         F_v, u, ubar, v, vbar, sigma, G, n)
