+++ demo/dolfin/stokes/hdg_linear_compressible_stokes.py
@@ -8,6 +8,7 @@ from dolfin import (
     DirichletBC, Form, MPI)
 
 import dolfin_dg.hdg_form
+import dolfin_dg.math
 
 k = 2
 n_eles = [8, 16, 32]
@@ -76,7 +77,7 @@ for run_no, n_ele in enumerate(n_eles):
     F = inner(F_v(rhou, grad(rhou), p), grad(v)) * dx - dot(gN, vbar) * dsN
 
     sigma = alpha / h
-    G = dolfin_dg.homogeneity_tensor(F_v, rhou)
+    G = dolfin_dg.math.homogeneity_tensor(F_v, rhou)
     hdg_term = dolfin_dg.hdg_form.HDGClassicalSecondOrder(
         F_v, rhou, rhoubar, v, vbar, sigma, G, n)
